refactor: extract delivery planning and state
